+++ Hostel.API/Controllers/PriceCategoryController.cs
@@ -78,6 +78,11 @@ namespace Hostel.API.Controllers
 
         public HttpResponseMessage Delete(HttpRequestMessage request, int id)
         {
+            PriceCategoryDTO data = service.Get(id);
+            if (data == null)
+            {
+                return request.CreateResponse(HttpStatusCode.NotFound);
+            }
             service.Delete(id);
             return request.CreateResponse(HttpStatusCode.OK);
         }
